+++ types/components/dv-cart.locale.d.ts
@@ -1,0 +1,28 @@
+declare namespace _default {
+    namespace en {
+        let empty: string;
+        let label: string;
+        let removeLabel: string;
+        let totalLabel: string;
+        let decreaseLabel: string;
+        let increaseLabel: string;
+        let quantityLabel: string;
+    }
+    namespace tr {
+        let empty_1: string;
+        export { empty_1 as empty };
+        let label_1: string;
+        export { label_1 as label };
+        let removeLabel_1: string;
+        export { removeLabel_1 as removeLabel };
+        let totalLabel_1: string;
+        export { totalLabel_1 as totalLabel };
+        let decreaseLabel_1: string;
+        export { decreaseLabel_1 as decreaseLabel };
+        let increaseLabel_1: string;
+        export { increaseLabel_1 as increaseLabel };
+        let quantityLabel_1: string;
+        export { quantityLabel_1 as quantityLabel };
+    }
+}
+export default _default;