+++ types/components/dv-confirm.locale.d.ts
@@ -1,0 +1,22 @@
+declare namespace _default {
+    namespace en {
+        let label: string;
+        let confirmingLabel: string;
+        let message: string;
+        let confirmLabel: string;
+        let cancelLabel: string;
+    }
+    namespace tr {
+        let label_1: string;
+        export { label_1 as label };
+        let confirmingLabel_1: string;
+        export { confirmingLabel_1 as confirmingLabel };
+        let message_1: string;
+        export { message_1 as message };
+        let confirmLabel_1: string;
+        export { confirmLabel_1 as confirmLabel };
+        let cancelLabel_1: string;
+        export { cancelLabel_1 as cancelLabel };
+    }
+}
+export default _default;