+++ types/components/dv-modal.locale.d.ts
@@ -1,0 +1,13 @@
+declare namespace _default {
+    namespace en {
+        let label: string;
+        let close: string;
+    }
+    namespace tr {
+        let label_1: string;
+        export { label_1 as label };
+        let close_1: string;
+        export { close_1 as close };
+    }
+}
+export default _default;